Update AOS.ahk
Tried to fix some undesireable things with the official client, unfortunatelly it's still not working properly.

Added a lot of new methods.
First of all, added a client check before every input since our own GUI was messing things up by getting the focus.

Also tried to fix high dpi active tab detection. Still have to test it better but seemed to work fine.

Added WaitTab() Method to wait for a tab if we want to. It's useful for some spells that are cast in inventory items like high alchemy and plank make for example.

Added methods to check if inventory is empty and/or inventory slot x is empty.

Added a method to click an inventory slot.

Added methods to check which spell book we are on and cast spell for standard and lunar magics.

Fixed gui size that was messed up with the last update. It's still not the way I want it since I want it to be scalable depending on DPI but works for now.
Made some changes to Input.SendAsyncInput()